+++ atlasdb-cassandra-multinode-tests/src/test/java/com/palantir/cassandra/multinode/TwoNodesDownTableManipulationTest.java
@@ -15,9 +15,9 @@
  */
 package com.palantir.cassandra.multinode;
 
-import org.apache.thrift.TException;
 import org.junit.Test;
 
+import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.ImmutableSet;
 import com.palantir.atlasdb.AtlasDbConstants;
 import com.palantir.atlasdb.keyvalue.api.TableReference;
@@ -31,7 +31,7 @@ public class TwoNodesDownTableManipulationTest extends AbstractDegradedClusterTe
     }
 
     @Test
-    public void createTableThrowsAndDoesNotChangeCassandraSchema() throws TException {
+    public void createTableThrowsAndDoesNotChangeCassandraSchema() {
         TableReference tableToCreate = TableReference.createWithEmptyNamespace("new_table");
         assertThrowsAtlasDbDependencyExceptionAndDoesNotChangeCassandraSchema(() ->
                 getTestKvs().createTable(tableToCreate, AtlasDbConstants.GENERIC_TABLE_METADATA));
@@ -39,13 +39,21 @@ public class TwoNodesDownTableManipulationTest extends AbstractDegradedClusterTe
     }
 
     @Test
-    public void dropTableThrowsAndDoesNotChangeCassandraSchema() throws TException {
+    public void createTablesThrowsAndDoesNotChangeCassandraSchema() {
+        TableReference tableToCreate = TableReference.createWithEmptyNamespace("new_table");
+        assertThrowsAtlasDbDependencyExceptionAndDoesNotChangeCassandraSchema(() ->
+                getTestKvs().createTables(ImmutableMap.of(tableToCreate, AtlasDbConstants.GENERIC_TABLE_METADATA)));
+
+    }
+
+    @Test
+    public void dropTableThrowsAndDoesNotChangeCassandraSchema() {
         assertThrowsAtlasDbDependencyExceptionAndDoesNotChangeCassandraSchema(() ->
                 getTestKvs().dropTable(TEST_TABLE));
     }
 
     @Test
-    public void dropTablesThrowsAndDoesNotChangeCassandraSchema() throws TException {
+    public void dropTablesThrowsAndDoesNotChangeCassandraSchema() {
         assertThrowsAtlasDbDependencyExceptionAndDoesNotChangeCassandraSchema(() ->
                 getTestKvs().dropTables(ImmutableSet.of(TEST_TABLE)));
     }